Renamed files for `JoltAreaImpl3D`
