+++ src/objects/jolt_area_impl_3d.cpp
@@ -1,4 +1,4 @@
-#include "jolt_area_3d.hpp"
+#include "jolt_area_impl_3d.hpp"
 
 #include "objects/jolt_body_3d.hpp"
 #include "spaces/jolt_broad_phase_layer.hpp"
