+++ src/servers/jolt_physics_server_3d.cpp
@@ -6,7 +6,7 @@
 #include "joints/jolt_joint_impl_3d.hpp"
 #include "joints/jolt_pin_joint_impl_3d.hpp"
 #include "joints/jolt_slider_joint_impl_3d.hpp"
-#include "objects/jolt_area_3d.hpp"
+#include "objects/jolt_area_impl_3d.hpp"
 #include "objects/jolt_body_3d.hpp"
 #include "servers/jolt_project_settings.hpp"
 #include "shapes/jolt_shape_3d.hpp"
