+++ src/spaces/jolt_contact_listener_3d.cpp
@@ -1,6 +1,6 @@
 #include "jolt_contact_listener_3d.hpp"
 
-#include "objects/jolt_area_3d.hpp"
+#include "objects/jolt_area_impl_3d.hpp"
 #include "objects/jolt_body_3d.hpp"
 #include "spaces/jolt_space_3d.hpp"
